Fixed all warnings resulted by -source 1.5 option
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@477070 13f79535-47bb-0310-9956-ffa450edef68

+++ integration-jmx/src/main/java/org/apache/mina/integration/jmx/IoServiceManager.java
@@ -28,7 +28,6 @@ import javax.management.ObjectName;
 
 import org.apache.mina.common.IoService;
 import org.apache.mina.common.IoSession;
-import org.apache.mina.management.IoSessionStat;
 import org.apache.mina.management.StatCollector;
 
 
@@ -124,25 +123,25 @@ public class IoServiceManager implements IoServiceManagerMBean, MBeanRegistratio
 
     public float getAverageByteReadThroughput()
     {
-        return collector.getBytesReadThroughput() / (float) collector.getSessionCount();
+        return collector.getBytesReadThroughput() / collector.getSessionCount();
     }
 
 
     public float getAverageByteWrittenThroughput()
     {
-        return collector.getBytesWrittenThroughput() / (float) collector.getSessionCount();    
+        return collector.getBytesWrittenThroughput() / collector.getSessionCount();    
     }
 
 
     public float getAverageMessageReadThroughput()
     {
-        return collector.getMsgReadThroughput() / (float) collector.getSessionCount();
+        return collector.getMsgReadThroughput() / collector.getSessionCount();
     }
 
 
     public float getAverageMessageWrittenThroughput()
     {
-        return collector.getMsgWrittenThroughput() / (float) collector.getSessionCount();    
+        return collector.getMsgWrittenThroughput() / collector.getSessionCount();    
     }
